## Summary

This PR adds a new `get_recent_changes` MCP tool and includes major performance optimizations that reduce test runtime by 98% (from 100s to 1.66s).

## Key Changes

### 🚀 New Feature: `get_recent_changes` Tool
- View recent commits from the GitHub repository
- **Filtering options:**
  - Time-based: Get changes from last N days
  - Count-based: Get last N commits (default: 10)
  - Path filtering: Filter by specific directories (_d/, _posts/, td/)
- **Performance:** Parallel API requests with semaphore limiting (max 15 concurrent)
- **Optional diff inclusion** for detailed change inspection

### ⚡ Performance Optimizations

#### 1. Fixed Critical Performance Bug in `read_blog_post`
- **Problem:** Downloaded ALL 200+ files to check for redirects on nonexistent URLs
- **Solution:** Use cached back-links.json metadata for redirect checking
- **Result:** 99% improvement (74s → 0.4s per request)

#### 2. Test Suite Optimization
- **Before:** 100+ seconds, frequent rate limit errors
- **After:** 1.66 seconds, no rate limits
- **Strategy:** Hybrid approach with real API sanity checks + mocked edge cases

### 🧹 Code Quality Improvements

- **DRY Refactoring:**
  - Added `get_blog_post_by_markdown_path()` helper
  - Added `format_blog_post()` formatter
  - Eliminated duplicate code patterns

- **Test Coverage:**
  - Comprehensive unit tests with mocks
  - Full E2E test suite
  - Performance regression tests

## Performance Metrics

| Metric | Before | After | Improvement |
|--------|--------|-------|-------------|
| Test Suite Runtime | 100+ seconds | 1.66 seconds | 98% faster |
| `read_blog_post` (not found) | 74 seconds | 0.4 seconds | 99% faster |
| GitHub API Calls (per test run) | ~200+ | ~20 | 90% reduction |

## Testing

All tests passing:
- ✅ 19 unit tests
- ✅ 28 E2E tests
- ✅ No rate limiting issues
- ✅ Performance benchmarks verified

## Breaking Changes

None - all existing functionality preserved.

🤖 Generated with [Claude Code](https://claude.ai/code)