@idvorkin-ai-tools 2025-09-14
## Response to Review

Thank you for the thorough review! I wanted to clarify the main concern:

### ✅ The `get_recent_changes` tool IS in the PR

The `get_recent_changes` tool is indeed implemented in this PR. You can see it:
- In blog_mcp_server.py at line 560 with the @mcp.tool decorator
- With comprehensive unit tests in test_unit.py
- With E2E tests in test_e2e.py

The tool supports:
- Time-based filtering (days parameter)
- Count-based filtering (commits parameter)  
- Path filtering
- Optional diff inclusion
- Parallel API requests with semaphore limiting (max 15 concurrent)

The PR diff shows 223 additions to blog_mcp_server.py which includes the full implementation of this new tool.

### Other Points Addressed

1. **Cache Duration**: Good suggestion about increasing from 5 to 30-60 minutes
2. **Redirect Checking**: The performance optimization for redirect checking is already implemented using cached back-links.json (lines 369-387 now use cached data)
3. **Error Handling**: Will consider more specific error messages in a follow-up
4. **Performance Benchmarks**: Great idea for future enhancement

The tool is functioning and can be tested with commands like:
```bash
just call get_recent_changes '{"commits": 5}'
just call get_recent_changes '{"days": 7, "path": "_d/"}'
```